+++ src/addon_generator/generators/__init__.py
@@ -1,9 +1,10 @@
 from .analytes_xml_generator import AddonXmlGenerationResult, generate_analytes_addon_xml
-from .protocol_generator import ProtocolJsonGenerationResult, generate_protocol_json
+from .protocol_json_generator import ProtocolJsonGenerationResult, ProtocolJsonGenerator, generate_protocol_json
 
 __all__ = [
     "AddonXmlGenerationResult",
     "ProtocolJsonGenerationResult",
     "generate_analytes_addon_xml",
     "generate_protocol_json",
+    "ProtocolJsonGenerator",
 ]